refactor(bp-train): migrate training pipeline to float64
Enable jax_enable_x64 in bp_train/__init__ (after the XLA device-count setup,
before the first jax import) and flip the explicit float32 dtype pins to
float64 so data, scales, model defaults, controls and batch padding all run in
float64.

- training_data.py, defaults.py, model_api.py, controls_store.py, harness.py:
  float32 -> float64.
- physical_solve.py: refresh the _EVENT_EPS comment for float64 (value
  unchanged; 1e-4 keeps ample headroom over float64 ULP at t~15).
- tests/test_model_api.py: expect float64 zero-width latent defaults.

+++ src/hybrax/train/controls_store.py
@@ -30,7 +30,7 @@ DEFAULT_RUNTIME_CONTROLS_CONFIG: dict[str, Any] = {
 }
 
 
-def _as_jax_array(values: Any, *, dtype: Any = jnp.float32) -> jax.Array:
+def _as_jax_array(values: Any, *, dtype: Any = jnp.float64) -> jax.Array:
     """Convert JSON-loaded values into a JAX array."""
     return jnp.asarray(values, dtype=dtype)
 
@@ -620,7 +620,7 @@ class ControlsStore(eqx.Module):
             bolus_event_Cin=(
                 jnp.zeros(
                     (len(process_order), 0, n_species),
-                    dtype=jnp.float32,
+                    dtype=jnp.float64,
                 )
                 if max_bolus_events == 0
                 else _as_jax_array(bolus_event_Cin_rows)
